Suggestions from CR
The number of lines changed is slightly higher than it might be because
I have also flipped the order of arguments in `gcd n m` to `gcd m n`.
This is to be compatible with `rat.cop` which became relevant in view of
the new lemma `add_circle.add_order_of_coe_rat`.