feat(analysis/convex/between): betweenness in a ring
Add some basic lemmas about betweenness in a ring as an affine space
over itself.

---

There is certainly plenty of scope for more lemmas in this area, but
the ones here are sufficient for what I want to do with them.  (That
next step - relating betweenness and affine combinations - also
depends on #17601, #17636, #17659 and #17660.)