feat(geometry/euclidean/circumcenter): reindexing
Add lemmas that reindexing a simplex along an `equiv` of index types
does not change the `circumsphere`, `circumcenter` or `circumradius`.

---

Regarding the comments in #17566 about reindexing along an embedding
rather than just an `equiv`, the lemmas here do need an `equiv` to get
all three equalities (which is thus why I was thinking of `equiv` when
writing that PR).  With an embedding you only get a weaker statement
about orthogonal projections (see the existing
`orthogonal_projection_circumcenter`).