feat(linear_algebra/affine_space/independent): spans of two affine combinations
Add some lemmas about spans of two points expressed as affine combinations.

---

Some of these are the sort of thing I suggested in #10632 as likely to
be useful intermediate steps towards Ceva's theorem, though that's not
what I'm aiming towards right now (rather, I'm aiming towards a
statement along the lines of: if lines from two vertices of a triangle
to interior points of the opposite sides meet, they do so inside the
triangle).