refactor(geometry/euclidean): split out spheres
Rearrange material related to spheres as follows:

* `geometry.euclidean.sphere.power` has most of the content from the
  previous `geometry.euclidean.sphere` (intersecting chords / secants,
  which at some point should probably be refactored using an explicit
  definition of the power of a point).

* `geometry.euclidean.sphere.ptolemy` has the proof of Ptolemy's
  theorem that was previously in `geometry.euclidean.sphere`.

* `geometry.euclidean.sphere.basic` has most of the material
  previously in `geometry.euclidean.basic`: definitions of `sphere`,
  `cospherical` and `concyclic` and associated lemmas.

* `geometry.euclidean.sphere.second_inter` has the definition and
  lemmas about `second_inter`.

There are no changes to API or proofs.

---

Questions for reviewers: much of the basic material about
`cospherical` and `sphere` is actually in a general metric space
context (it doesn't use the space of vectors at all).  Should that
material move to a file under `topology.metric_space` (and associated
namespace different from `euclidean_geometry`)?  If so, what should
the bundled `sphere` type be called in such a file and namespace
(there's an existing unbundled `metric.sphere`, so it can't be called
that)?  If it does move there, presumably some other pieces should
eventually move as well (for example, `concyclic` is more generally
meaningful in a non-Euclidean `normed_add_torsor`).

If the material should move under `topology.metric_space`, I still
think splitting it out of `geometry.euclidean.basic` is a reasonable
start and such a move under `topology.metric_space` would best be done
in a separate PR.