chore: Add toml file for archive and counterexamples (#18388)
The end goal here is to surface these in doc-gen.

For doc-gen to behave these need to be mutually importable. A step is added to the "test" run that verifies this by trying to import everything at the same time.

For this new test to pass, some files need namespaces wrapped around their content.  Probably we should be more principled about this namespace wrapping, but that can be left as future work, as all we really care about to start integrating with doc-gen is avoiding clashes.