chore(algebra/algebra/basic): missing lemmas about the `↑` spelling of `algebra_map`
This only really scratches the surface; in general we can't afford to have two different spellings here as it forces us to duplicate every result about `ring_hom`.
A previously refactor (#17048) made these lemmas unecessary, but it was reverted to aid with the port.
Since these lemmas are needed in mathlib3 PRs, lets have them anyway.